Reorganizando la estructura de la API
