+++ src/main/java/com/example/apiexportexceltoemail/dto/ReportDTO.java
@@ -1,0 +1,17 @@
+package com.example.apiexportexceltoemail.dto;
+
+import java.time.LocalDate;
+
+public interface report {
+
+    private LocalDate ExecuteDate;
+    private String errorMessage;
+    private String order;
+    private String item;
+    private String type;
+    private String sapCode;
+    private String fileName;
+    private String status;
+
+
+}
