+++ src/main/java/com/example/apiexportexceltoemail/persistences/models/ParamFecha.java
@@ -1,0 +1,2 @@
+package com.example.apiexportexceltoemail.persistences.models.entities;public class paramFecha {
+}
