--- src/test/java/com/example/apiexportexceltoemail/TestConexionSQL.java
@@ -1,30 +1,0 @@
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-import org.junit.jupiter.api.Test;
-
-public class TestConexionSQL {
-
-    @Test
-    public void testConexion() {
-        Connection conn = null;
-        try {
-            conn = DriverManager.getConnection("jdbc:sqlserver://localhost:1433;databaseName=DBName;user=UserName;password=Password");
-            assertTrue(conn != null);
-        } catch (SQLException ex) {
-            ex.printStackTrace();
-        } finally {
-            if (conn != null) {
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-
-}
